requirements.txt
absl-py
aiohttp
alembic==1.7.7
antlr4-python3-runtime
async-timeout==3.0.1
attrs
autopage==0.5.0
backcall
backports.functools-lru-cache
beautifulsoup4==4.11.1
blinker==1.4
boto==2.49.0
boto3
botocore
bravado
bravado-core==5.17.0
brotlipy==0.7.0
cachetools
certifi
cffi
chardet==3.0.4
charset-normalizer
click
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorama
colorlog==6.6.0
cryptography
cycler
Cython
dataclasses
debugpy
decorator==4.4.2
dgl==0.9.0
entrypoints
filelock==3.8.0
fonttools==4.29.1
future
gdown==4.5.3
gensim
gitdb
GitPython
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
googledrivedownloader
greenlet==1.1.2
grpcio
humanize
idna
importlib-metadata
importlib-resources
ipykernel
ipython
isodate==0.6.1
jedi
Jinja2
jmespath
joblib
jsonref==0.2
jsonschema
jupyter-client
jupyter-core
kiwisolver
littleutils==0.2.2
Mako==1.2.0
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
monotonic==1.5
msgpack
multidict
munkres==1.1.4
neptune-client
nest-asyncio
networkx
numpy
nvidia-ml-py==11.515.75
nvitop==0.11.0
oauthlib
ogb==1.3.5
olefile
omegaconf
optuna==2.10.0
outdated==0.2.2
packaging
pandas==1.2.3
parso
patsy==0.5.2
pbr==5.8.1
pexpect
pickleshare
Pillow==8.4.0
ply==3.11
prettytable==3.2.0
prompt-toolkit
protobuf==3.13.0
psutil
ptyprocess
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
PyGCL==0.1.1
Pygments
PyJWT
pyOpenSSL
pyparsing
pyperclip==1.8.2
PyQt5-sip==12.11.0
pyrsistent
PySocks
python-dateutil
python-louvain
python-magic==0.4.27
pytz
PyYAML==5.4.1
pyzmq==19.0.2
rdflib==6.1.1
requests
requests-oauthlib
rsa
s3transfer
scikit-learn
scipy
seaborn
simplejson
sip
six
smart-open
smmap
soupsieve==2.3.2.post1
SQLAlchemy==1.4.32
statsmodels==0.13.2
stevedore==3.5.0
swagger-spec-validator
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor==2.2.0
threadpoolctl
toml
torch==1.10.2
torch-cluster
torch-geometric
torch-scatter
torch-sparse
torch-spline-conv
torchaudio==0.10.2
torchvision==0.11.3
tornado
tqdm
traitlets
typing-extensions
urllib3
wcwidth
websocket-client
Werkzeug
yacs==0.1.6
yarl
zipp